Bug 971483 - [System2] Instantiable DevToolsLayer r=alive
