+++ apps/system/index.html
@@ -314,8 +314,8 @@
     <script defer src="js/layout_manager.js"></script>
 
     <!-- Devtools -->
-    <link rel="stylesheet" type="text/css" href="style/devtools_layers.css">
-    <script defer src="js/devtools_layers.js"></script>
+    <link rel="stylesheet" type="text/css" href="style/devtools_view.css">
+    <script defer src="js/devtools_view.js"></script>
 
     <!-- Edge Gestures -->
     <script defer src="js/stack_manager.js"></script>
