+++ apps/system/test/unit/bootstrap_test.js
@@ -10,6 +10,7 @@ requireApp('system/shared/test/unit/mocks/mock_navigator_moz_settings.js');
 requireApp('system/shared/test/unit/mocks/mock_settings_listener.js');
 requireApp('system/shared/test/unit/mocks/mock_settings_url.js');
 requireApp('system/test/unit/mock_applications.js');
+requireApp('system/test/unit/mock_devtools_view.js');
 requireApp('system/test/unit/mock_ftu_launcher.js');
 requireApp('system/test/unit/mock_homescreen_launcher.js');
 requireApp('system/test/unit/mock_places.js');
@@ -24,11 +25,13 @@ mocha.globals([
   'cancelHomeTouchend',
   'secureWindowManager',
   'secureWindowFactory',
+  'devtoolsView',
   'ttlView',
 ]);
 
 var mocksForBootstrap = new MocksHelper([
   'Applications',
+  'DevtoolsView',
   'FtuLauncher',
   'HomescreenLauncher',
   'Places',
