+++ apps/system/test/unit/mock_devtools_view.js
@@ -1,0 +1,6 @@
+/* exported MockDevtoolsView */
+
+'use strict';
+
+function MockDevtoolsView() {
+}
